Change expected issue format to be less visually jarring
This will parse issue labels in the "Estimate: %number" format so we can have an actually readable issue label on the other pages.

An option to configure this would be preferable, but I've literally no idea how to create Chrome extension UIs and I don't care to learn more about it right now. Maybe later?